+++ src/app/om_afl.h
@@ -16,24 +16,24 @@ typedef struct {
 } om_afl_t;
 
 typedef struct {
-  size_t offset;
-  size_t scope;
+  uint32_t offset;
+  uint32_t scope;
   void* template;
 } _om_afl_filter_list_t;
 
 typedef struct {
-  size_t offset;
+  uint32_t offset;
   uint32_t start;
   uint32_t range;
 } _om_afl_filter_range_t;
 
 typedef struct {
-  size_t offset;
-  size_t size;
+  uint32_t offset;
+  uint32_t size;
 } _om_afl_filter_decompose_t;
 
 typedef struct {
-  size_t length;
+  uint32_t length;
   OM_AFL_MODE_t mode;
   union {
     _om_afl_filter_list_t list;
@@ -50,9 +50,9 @@ om_filter_t* om_afl_filter_create(om_topic_t* target);
 
 om_status_t om_afl_add_filter(om_afl_t* afl, om_filter_t* filter);
 
-om_status_t om_afl_set_filter(om_filter_t* filter, uint8_t mode, size_t offset,
-                              size_t length, size_t scope, uint32_t arg,
-                              void* template);
+om_status_t om_afl_set_filter(om_filter_t* filter, uint8_t mode,
+                              uint32_t offset, uint32_t length, uint32_t scope,
+                              uint32_t arg, void* template);
 
 om_status_t _om_afl_filter_check(om_filter_t* filter, om_msg_t* msg);
